test: reclassify misclassified integration tests (#141)
Move pure unit tests (no external deps) from tests/integration/ to
tests/unit/, port unique integration-only assertions into existing unit
files, and split tests/integration/monsterUpload.test.ts into focused
tests/unit/monster-upload/ sub-files. Delete tests proven redundant.

Collapse jest.docker.config.js into jest.integration.config.js: remove
testPathIgnorePatterns and test:docker script — the split only existed
because mocked tests didn't need Docker. Integration suite now runs all
real-deps tests under a single config with no exclusions.
