+++ openspec/changes/cleanup-misclassified-integration-tests/.openspec.yaml
@@ -1,0 +1,2 @@
+schema: sdd-with-feedback-loop
+created: 2026-04-15
