+++ tests/unit/validation/password.test.ts
